requirements.txt
Flask
requests
openai
python-dotenv
gunicorn